+++ tests/unit/config/test_bracket_pos_args.py
@@ -177,13 +177,3 @@ def test_no_posargs_inside_substitutions(get_option, monkeypatch):
         ["bar"],
     )
     assert got == ["foo", "[]"]
-
-
-@pytest.fixture
-def get_option(newconfig):
-    def do(tox_ini, option_name, pos_args=()):
-        config = newconfig(list(pos_args), tox_ini)
-        print(type(config.envconfigs), config.envconfigs.keys())
-        return getattr(config.envconfigs["python"], option_name)
-
-    return do
